# docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/actors/{address}/proposals": {
      "get": {
        "operationId": "ActorProposalsController_list",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-delimited sort fields (prefix with - for desc)",
            "schema": {
              "example": "slug,-created_at",
              "type": "string"
            }
          },
          {
            "name": "dao",
            "required": false,
            "in": "query",
            "description": "Comma-delimited DAO slugs",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Comma-delimited proposal states",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "proposer",
            "required": false,
            "in": "query",
            "description": "0x-prefixed proposer address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "binding",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "voting_starts_at_min",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "voting_starts_at_max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorProposalListResponseDto"
                }
              }
            }
          },
          "301": {
            "description": "Redirect to canonical actor address"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["actors"]
      }
    },
    "/v1/actors/{address}/votes": {
      "get": {
        "operationId": "ActorVotesController_list",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-delimited sort fields (prefix with - for desc)",
            "schema": {
              "example": "slug,-created_at",
              "type": "string"
            }
          },
          {
            "name": "dao",
            "required": false,
            "in": "query",
            "description": "Comma-delimited DAO slugs",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Comma-delimited proposal states",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "proposer",
            "required": false,
            "in": "query",
            "description": "0x-prefixed proposer address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "binding",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "voting_starts_at_min",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "voting_starts_at_max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorVoteListResponseDto"
                }
              }
            }
          },
          "301": {
            "description": "Redirect to canonical actor address"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["actors"]
      }
    },
...
    "/v1/actors/{address}/analytics/cross-dao": {
      "get": {
        "operationId": "ActorAnalyticsController_crossDao",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrossDaoActorDto"
                }
              }
            }
          },
          "301": {
            "description": "Redirect to canonical actor address"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["analytics"]
      }
    },
...
  },
  "info": {
    "title": "Kvorum API",
    "description": "Kvorum M1 DAO + proposal read API",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "type": "http",
        "description": "ADR-025 API-key bearer authentication"
      }
    },
    "schemas": {
...
      "ProblemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "violations": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": ["type", "title", "status", "detail", "instance"]
      },
      "ActorProposalMetaDto": {
        "type": "object",
        "properties": {
          "confirmed": {
            "type": "boolean"
          }
        },
        "required": ["confirmed"]
      },
      "ActorProposalListItemDto": {
        "type": "object",
        "properties": {
          "proposal_id": {
            "type": "string"
          },
          "source_type": {
            "type": "string"
          },
          "dao_slug": {
            "type": "string"
          },
          "title": {
            "type": "object",
            "nullable": true
          },
          "state": {
            "type": "string"
          },
          "voting_starts_at": {
            "type": "object",
            "nullable": true
          },
          "voting_ends_at": {
            "type": "object",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "_meta": {
            "$ref": "#/components/schemas/ActorProposalMetaDto"
          }
        },
        "required": [
          "proposal_id",
          "source_type",
          "dao_slug",
          "state",
          "voting_starts_at",
          "voting_ends_at",
          "created_at",
          "_meta"
        ]
      },
      "PaginationDto": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number"
          },
          "next_cursor": {
            "type": "object",
            "nullable": true
          },
          "has_more": {
            "type": "boolean"
          }
        },
        "required": ["limit", "next_cursor", "has_more"]
      },
      "ActorProposalListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActorProposalListItemDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": ["data", "pagination"]
      },
      "EmbeddedProposalLinksDto": {
        "type": "object",
        "properties": {
          "proposal": {
            "type": "string"
          }
        },
        "required": ["proposal"]
      },
      "EmbeddedProposalMetaDto": {
        "type": "object",
        "properties": {
          "links": {
            "$ref": "#/components/schemas/EmbeddedProposalLinksDto"
          }
        },
        "required": ["links"]
      },
      "EmbeddedProposalDto": {
        "type": "object",
        "properties": {
          "proposal_id": {
            "type": "string"
          },
          "source_type": {
            "type": "string"
          },
          "dao_slug": {
            "type": "string"
          },
          "title": {
            "type": "object",
            "nullable": true
          },
          "state": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "voting_ends_at": {
            "type": "object",
            "nullable": true
          },
          "_meta": {
            "$ref": "#/components/schemas/EmbeddedProposalMetaDto"
          }
        },
        "required": ["proposal_id", "source_type", "dao_slug", "state", "created_at", "_meta"]
      },
      "ActorVoteMetaDto": {
        "type": "object",
        "properties": {
          "confirmed": {
            "type": "boolean"
          }
        },
        "required": ["confirmed"]
      },
      "ActorVoteListItemDto": {
        "type": "object",
        "properties": {
          "vote_id": {
            "type": "string"
          },
          "proposal": {
            "$ref": "#/components/schemas/EmbeddedProposalDto"
          },
          "primary_choice": {
            "type": "object",
            "nullable": true
          },
          "voting_power_reported": {
            "type": "string"
          },
          "cast_at": {
            "type": "object",
            "nullable": true
          },
          "_meta": {
            "$ref": "#/components/schemas/ActorVoteMetaDto"
          }
        },
        "required": ["vote_id", "proposal", "voting_power_reported", "cast_at", "_meta"]
      },
      "ActorVoteListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActorVoteListItemDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": ["data", "pagination"]
      },
...
      "CrossDaoSummaryDto": {
        "type": "object",
        "properties": {
          "dao_slug": {
            "type": "string"
          },
          "votes_cast": {
            "type": "number"
          },
          "proposals_proposed": {
            "type": "number"
          },
          "current_voting_power": {
            "type": "string"
          },
          "last_active_at": {
            "type": "object",
            "nullable": true
          },
          "alignment_with_majority_pct": {
            "type": "object",
            "nullable": true
          }
        },
        "required": ["dao_slug", "votes_cast", "proposals_proposed", "current_voting_power"]
      },
      "AnalyticsMetaDto": {
        "type": "object",
        "properties": {
          "confirmed": {
            "type": "boolean"
          },
          "derived_through": {
            "type": "object",
            "nullable": true
          }
        },
        "required": ["confirmed"]
      },
      "CrossDaoActorDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "actor_id": {
            "type": "string"
          },
          "daos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CrossDaoSummaryDto"
            }
          },
          "_meta": {
            "$ref": "#/components/schemas/AnalyticsMetaDto"
          }
        },
        "required": ["address", "actor_id", "daos", "_meta"]
      },
...
    }
  }
}